backend: Modify the function of create namespace location
