+++ src/backend/resources/namespace/list.go
@@ -2,7 +2,6 @@ package namespace
 
 import (
 	"k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
@@ -32,25 +31,6 @@ func GetNamespace(cli *kubernetes.Clientset, namespace string) (*v1.Namespace, e
 		if ns.Name == namespace {
 			return &ns, nil
 		}
-	}
-	return nil, nil
-}
-
-func UpdateNamespace(cli *kubernetes.Clientset, ns *v1.Namespace) (*v1.Namespace, error) {
-	newNS, err := cli.CoreV1().Namespaces().Update(ns)
-	if err != nil {
-		return nil, err
-	}
-	return newNS, nil
-}
-
-func CreateNotExitNamespace(cli *kubernetes.Clientset, ns *v1.Namespace) (*v1.Namespace, error) {
-	_, err := cli.CoreV1().Namespaces().Get(ns.Name, metaV1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			return cli.CoreV1().Namespaces().Create(ns)
-		}
-		return nil, err
 	}
 	return nil, nil
 }
