+++ src/backend/resources/namespace/namespace.go
@@ -3,9 +3,30 @@ package namespace
 import (
 	"github.com/Qihoo360/wayne/src/backend/resources/common"
 	"github.com/Qihoo360/wayne/src/backend/util"
+	"k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
+
+func UpdateNamespace(cli *kubernetes.Clientset, ns *v1.Namespace) (*v1.Namespace, error) {
+	newNS, err := cli.CoreV1().Namespaces().Update(ns)
+	if err != nil {
+		return nil, err
+	}
+	return newNS, nil
+}
+
+func CreateNotExitNamespace(cli *kubernetes.Clientset, ns *v1.Namespace) (*v1.Namespace, error) {
+	_, err := cli.CoreV1().Namespaces().Get(ns.Name, metaV1.GetOptions{})
+	if err != nil {
+		if errors.IsNotFound(err) {
+			return cli.CoreV1().Namespaces().Create(ns)
+		}
+		return nil, err
+	}
+	return nil, nil
+}
 
 func ResourcesUsageByNamespace(cli *kubernetes.Clientset, namespace, selector string) (*common.ResourceList, error) {
 	podList, err := cli.CoreV1().Pods(namespace).List(metaV1.ListOptions{
